general refactor of parsing, operation refactor, while implementation
